refactor: 💡 replace deepClone() usage by json-clone
