+++ src/json-patch/applyPatch/v3.ts
@@ -1,4 +1,4 @@
-import {deepClone} from '../util';
+import {clone as deepClone} from '../../json-clone/clone';
 import {Operation} from '../types';
 import {findByPointer, unescapeComponent} from '../../json-pointer';
 import {deepEqual} from '../../json-equal/deepEqual';
